% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelHMMR.R
\docType{class}
\name{ModelHMMR-class}
\alias{ModelHMMR-class}
\alias{ModelHMMR}
\title{A Reference Class which represents a fitted HMMR model.}
\description{
ModelHMMR represents a \link[=ModelHMMR]{HMMR} model for which parameters have
been estimated.
}
\section{Fields}{

\describe{
\item{\code{paramHMMR}}{A \link{ParamHMMR} object. It contains the estimated values of the parameters.}

\item{\code{statHMMR}}{A \link{StatHMMR} object. It contains all the statistics associated to the HMMR model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("predicted", "filtered", "smoothed", "regressors"))}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item "predicted"
      \item "filtered"
      \item "smoothed"
      \item "regressors"
    }
  }
}}

\item{\code{summary()}}{Summary method.}
}}

\seealso{
\link{ParamHMMR}, \link{StatHMMR}
}
