% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamHMMR.R
\docType{class}
\name{ParamHMMR-class}
\alias{ParamHMMR-class}
\alias{ParamHMMR}
\title{A Reference Class which contains parameters of a HMMR model.}
\description{
ParamHMMR contains all the parameters of a HMMR model.
}
\section{Fields}{

\describe{
\item{\code{X}}{Numeric vector of length \emph{m} representing the covariates/inputs
\eqn{x_{1},\dots,x_{m}}.}

\item{\code{Y}}{Numeric vector of length \emph{m} representing the observed
response/output \eqn{y_{1},\dots,y_{m}}.}

\item{\code{m}}{Numeric. Length of the response/output vector \code{Y}.}

\item{\code{K}}{The number of regimes (mixture components).}

\item{\code{p}}{The order of the polynomial regression.}

\item{\code{variance_type}}{Character indicating if the model is homoskedastic
(\code{variance_type} = "homoskedastic") or heteroskedastic
(\code{variance_type} = "heteroskedastic").}

\item{\code{prior}}{The prior probabilities of the Markov chain.}

\item{\code{trans_mat}}{The transition matrix of the Markov chain.}

\item{\code{beta}}{Parameters of the polynomial regressions.
\eqn{\beta = (\beta_{1},\dots,\beta_{K})} is a matrix of dimension
\eqn{(p + 1, K)}, with \emph{p} the order of the polynomial regression.}

\item{\code{sigma2}}{The variances for the \emph{K} regimes. If HMMR model is
homoskedastic (\code{variance_type} = "homoskedastic") then \code{sigma2} is a
matrix of size \eqn{(1, 1)}, else if HMMR model is heteroskedastic
(\code{variance_type} = "heteroskedastic") then \code{sigma2} is a matrix of size
\eqn{(K, 1)}.}

\item{\code{nu}}{The degree of freedom of the HMMR model.}

\item{\code{phi}}{A designed matrix for the polynomial regressions.}
}}

\section{Methods}{

\describe{
\item{\code{initParam(try_algo = 1)}}{Method to initialize parameters \code{prior}, \code{trans_mat},
      \code{beta} and \code{sigma2}.

      If try_algo = 1 then \code{beta} and \code{sigma2} are
      initialized by segmenting uniformly into \code{K} contiguous segments
      the response Y. Otherwise, \code{beta} and \code{sigma2} are
      initialized by segmenting randomly the response Y into \code{K} segments.}

\item{\code{MStep(statHMMR)}}{Method used in the EM algorithm to learn the parameters of the HMMR model
based on statistics provided by \code{statHMMR}.}
}}

