% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{selectHMMR}
\alias{selectHMMR}
\title{selectHMMR is used to select a HMMR model based on criteria such as BIC or AIC.}
\usage{
selectHMMR(X, Y, Kmin = 1, Kmax = 10, pmin = 0, pmax = 4,
  criterion = c("BIC", "AIC"))
}
\arguments{
\item{X}{Numeric vector of length \emph{m} representing the covariates.}

\item{Y}{Matrix of size \eqn{(n, m)} representing \emph{n} functions of \code{X}
observed at points \eqn{1,\dots,m}.}

\item{Kmin}{The minimum number of regimes (mixture components).}

\item{Kmax}{The maximum number of regimes (mixture components).}

\item{pmin}{The minimum order of the polynomial regression.}

\item{pmax}{The maximum order of the polynomial regression.}

\item{criterion}{The criterion used to select the "best" HMMR model.}
}
\value{
selectHMMR returns an object of class \link{ModelHMMR}
representing the "best" HMMR model according to the selected \code{criterion}.
}
\description{
selectHMMR is used to select a HMMR model based on criteria such as BIC or AIC.
}
\details{
selectHMMR is used to select the "best" HMMR model according to some
criteria such as BIC or AIC. This function runs every HMMR model by varying
the number of regimes \code{K} from \code{Kmin} to \code{Kmax} and the order of the
polynomial regression \code{p} from \code{pmin} to \code{pmax}.
}
\seealso{
\link{ModelHMMR}
}
