% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelHMMR.R
\docType{class}
\name{ModelHMMR-class}
\alias{ModelHMMR-class}
\alias{ModelHMMR}
\title{A Reference Class which represents a fitted HMMR model.}
\description{
ModelHMMR represents an estimated HMMR model.
}
\section{Fields}{

\describe{
\item{\code{paramHMMR}}{A \link{ParamHMMR} object. It contains the estimated
values of the parameters.}

\item{\code{statHMMR}}{A \link{StatHMMR} object. It contains all the statistics
associated to the HMMR model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("predicted", "filtered", "smoothed", "regressors"))}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item \code{"predicted" = } Predicted time series and predicted
      regime probabilities.
      \item \code{"filtered" = } Filtered time series and filtering
      regime probabilities.
      \item \code{"smoothed" = } Smoothed time series, and
      segmentation.
      \item \code{"regressors" = } Polynomial regression components.
    }
  }
}
By default, all the above graphs are produced.}

\item{\code{summary()}}{Summary method.}
}}

\seealso{
\link{ParamHMMR}, \link{StatHMMR}
}
